Eliminated the 'mptk.h: left shift count >= width of type' warnings.
git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/mptk/trunk@590 0f77665a-23fb-0310-bc23-e91fd8efc3e1

+++ src/libmptk/mp_types.h
@@ -104,16 +104,16 @@ typedef struct {
 /* Bounds of integer types */
 
 const unsigned short int MP_MAX_BITS_SHORT_INT = 8*sizeof(short int) - 1; /* 15 */
-const short int MP_MIN_SHORT_INT = (short int) (1<<MP_MAX_BITS_SHORT_INT); /*-32768*/
-const short int MP_MAX_SHORT_INT = (short int) ((1<<MP_MAX_BITS_SHORT_INT) - 1); /*32767*/
+const short int MP_MIN_SHORT_INT = (short int)   ( 1 << MP_MAX_BITS_SHORT_INT ); /*-32768*/
+const short int MP_MAX_SHORT_INT = (short int) ( ( 1 << MP_MAX_BITS_SHORT_INT ) - 1); /*32767*/
 
 const unsigned short int MP_MAX_BITS_UNSIGNED_SHORT_INT = 8*sizeof(unsigned short int); /* 16 */
 const unsigned short int MP_MIN_UNSIGNED_SHORT_INT = 0;
 const unsigned short int MP_MAX_UNSIGNED_SHORT_INT = (unsigned short int) (-1); /* 65535 */
 
 const unsigned short int MP_MAX_BITS_LONG_INT = 8*sizeof(long int) - 1; /* 31 */
-const long int MP_MIN_LONG_INT = (long int) (1<<MP_MAX_BITS_LONG_INT); /*-2147483648*/
-const long int MP_MAX_LONG_INT = (long int) ((1<<MP_MAX_BITS_LONG_INT) - 1); /* 2147483647 */
+const long int MP_MIN_LONG_INT = (long int)   ( (long)(1) << MP_MAX_BITS_LONG_INT ); /*-2147483648*/
+const long int MP_MAX_LONG_INT = (long int) ( ( (long)(1) << MP_MAX_BITS_LONG_INT ) - (long)(1) ); /* 2147483647 */
 
 const unsigned short int MP_MAX_BITS_UNSIGNED_LONG_INT = 8*sizeof(unsigned long int); /* 32 */
 const unsigned long int MP_MIN_UNSIGNED_LONG_INT = 0;
